Refactor Markdown class component to functional component
